created new project with MapVisual files for creating the bmp
Currently its being tested as a new project, when used we should just copy the MapVisual.c and .h files to the Utils folder and then call them as needed.

Code is not yet optimised but it works. Next steps are trying to create function for just one location and optimizing the code so it doesn't have warnings.

Dont know how it will function will very different map sizes