Added movment functions
The functions are at the start before init(). Take a look in TestMode3 to see how to use iterations.

We modified the speeds to take only iteration so its a bit fast but that can easily be changed. TODO : figure out how iterations work and possibly change to a while loop.

Co-Authored-By: jsha485 <68843007+jsha485@users.noreply.github.com>